Add integration test using TFL-Microspeech example. (#21)
This PR adds an integration test using example project from
https://github.com/ARM-software/AVH-TFLmicrospeech.
The example project is pulled in, patched to use the latest generated
pack, and built/executed on Arm Virtual Hardware.

Once this workflow is merged to main it will be triggered whenever the
build workflow completes. The pack generated during build workflow will
then be tested automatically. If the build workflow runs as part as a PR
check the result of this "integration test" is added to the results of
the PR. This can help to check that a PR doesn't break downstream
projects.