Explicitly cast (void *) to (const int32_t *)
An implicit cast right there is the only reason CMSIS-NN can't
be compiled with a C++ compiler.

I am aware this is a C project and not a C++ one but for
my specific usecase I need to unfortunately compile it
with a c++ compiler. Adding this patch to my fork works,
and since it's a minor change that works in C just fine
I wondered if you'd want this commit upstream.

Thanks for the great work!